refactor(runtime): remove http_adapter module and inline request/response handling
