--- crates/rari/src/runtime/http_adapter/mod.rs
@@ -1,3 +1,0 @@
-pub mod adapter;
-
-pub use adapter::HttpAdapter;
